usuarios/backend/requirements.txt
flask
flask-cors
ecdsa
werkzeug
psycopg2-binary
